adding the login dialog and cleaning a bit the code
