+++ app/lib/browser/index.js
@@ -6,7 +6,7 @@
   const trayNotifications = require('./tray-notifications');
   const nativeNotifications = require('./native-notifications');
   require('./zoom')();
-  
+
   const iconPath = path.join(__dirname, '../assets/icons/icon-96x96.png');
 
   trayNotifications({
