+++ app/lib/browser/native-notifications.js
@@ -5,7 +5,7 @@ const { nativeImage } = require('electron');
 exports = module.exports = ({ ipc, iconPath }) => {
   return () => {
     const icon = nativeImage.createFromPath(iconPath);
-    if (typeof Notify !== 'undefined') {      
+    if (typeof Notify !== 'undefined') {
       Notify.prototype.show = function show() {
         const notification = new Notification(this.title, {
           body: this.options.body,
