+++ app/lib/index.js
@@ -1,24 +1,17 @@
 'use strict';
-
 const windowStateKeeper = require('electron-window-state');
 const path = require('path');
-const { shell, app, ipcMain, remote, BrowserWindow } = require('electron');
+const { shell, app, ipcMain, BrowserWindow } = require('electron');
 const configBuilder = require('./config');
+const login = require('./login');
 const NativeNotification = require('electron-native-notification');
-
-const DEFAULT_WINDOW_WIDTH = 0;
-const DEFAULT_WINDOW_HEIGHT = 0;
-
 const Menus = require('./menus');
-
-let menus;
-
 
 function createWindow(iconPath) {
   // Load the previous state with fallback to defaults
   let windowState = windowStateKeeper({
-    defaultWidth: DEFAULT_WINDOW_WIDTH,
-    defaultHeight: DEFAULT_WINDOW_HEIGHT
+    defaultWidth: 0,
+    defaultHeight: 0
   });
 
   // Create the window
@@ -52,8 +45,8 @@ function createWindow(iconPath) {
   return window;
 }
 
-app.commandLine.appendSwitch('auth-server-whitelist','*');
-app.commandLine.appendSwitch('enable-ntlm-v2','true');
+app.commandLine.appendSwitch('auth-server-whitelist', '*');
+app.commandLine.appendSwitch('enable-ntlm-v2', 'true');
 
 app.on('ready', () => {
 
@@ -61,10 +54,9 @@ app.on('ready', () => {
     app.getAppPath(),
     'lib/assets/icons/icon-96x96.png'
   );
-  const window = createWindow(iconPath);
+  var window = createWindow(iconPath);
   const config = configBuilder(app.getPath('userData'));
-
-  menus = new Menus(config, iconPath);
+  let menus = new Menus(config, iconPath);
   menus.register(window);
 
   window.on('page-title-updated', (event, title) =>
@@ -78,17 +70,17 @@ app.on('ready', () => {
   });
 
   ipcMain.on('notifications', async (e, msg) => {
-    if (msg.count>0){  
-      const body = "You got " + msg.count + " notification(s). " + ((msg.text) ? " <i>" + msg.text + "</i> " : "");    
+    if (msg.count > 0) {
+      const body = "You got " + msg.count + " notification(s). " + ((msg.text) ? " <i>" + msg.text + "</i> " : "");
       const notification = new NativeNotification(
-        "Microsoft Teams", 
+        "Microsoft Teams",
         {
           "body": body,
           "icon": iconPath,
-         });
+        });
       if (notification.show !== undefined) {
         notification.show();
-      } 
+      }
     }
   });
 
@@ -99,7 +91,7 @@ app.on('ready', () => {
 
   window.webContents.on('login', (event, request, authInfo, callback) => {
     event.preventDefault();
-    loginService(callback);
+    login.loginService(callback);
   });
 
   if (config.userAgent === 'edge') {
@@ -107,19 +99,19 @@ app.on('ready', () => {
   } else {
     window.webContents.setUserAgent(config.chromeUserAgent);
   }
-  
+
   window.once('ready-to-show', () => window.show());
 
-  window.webContents.on('did-finish-load', function() {
+  window.webContents.on('did-finish-load', function () {
     window.webContents.insertCSS('#download-mobile-app-button, #download-app-button, #get-app-button { display:none; }');
     window.webContents.insertCSS('.zoetrope { animation-iteration-count: 1 !important; }');
   });
 
   window.on('closed', () => window = null);
 
-  window.loadURL(config.url)
-  // loginService((username, password) => {
-  //   window.loadURL(config.url)
+  window.loadURL(config.url);
+  // login.loginService((username, password) => {
+  //   window.loadURL(config.url);
   // });
 });
 
@@ -127,30 +119,4 @@ app.on('login', function (event, webContents, request, authInfo, callback) {
   if (typeof config !== 'undefined' && typeof config.firewallUsername !== 'undefined') {
     callback(config.firewallUsername, config.firewallPassword);
   }
-  
 });
-
-
-function loginService(callback) {
-  const win = new BrowserWindow({
-    width: 363,
-    height: 124,
-    frame: false,
-    preload: path.join(__dirname, 'login', 'index.js'),
-
-    show: false,
-    autoHideMenuBar: true,
-  });
-
-  win.once('ready-to-show', () => {
-    win.show();
-  });
-
-  ipcMain.on('submitForm', function(event, data) {
-    console.log('submitForm called', data);
-    callback(data.username, data.password);
-    win.close();
-  });
-
-  win.loadURL(`file://${__dirname}/login/login.html`);
-}
