+++ app/lib/menus/index.js
@@ -1,5 +1,4 @@
 'use strict';
-
 const {
   app,
   nativeImage,
